+++ CHANGELOG.md
@@ -118,6 +118,7 @@
 3. Updated Cpp2IL to v2021.2.4.
 4. Fixed Issue with Assembly Generator not responding to RemoteAPI when Forcing Dumper Version.
 5. Fixed Issue with Offline Mode throwing an NRE during Assembly Generation.
+6. Fixed Issue with Launch Options causing Crash.
 
 ---