+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 9. Fixed Issue with Older OS Check running under Wine/Proton.
 10. [StartScreen] Resize fix x86.    (Credits to [Slaynash](https://github.com/Slaynash) :3)
 11. [StartScreen] Fixed crash on mono.    (Credits to [Slaynash](https://github.com/Slaynash) :3)
+12. Rewrote MelonLaunchOptions Loading.
 
 ---